Added "useTopWindow"boolean
A change which adds the option of showing the toast notification on the top window or the window of the current root controller.

default  useTopWindow = true

Useful if we don't want the toast attaching to any modal view controllers which may pop up. This doesn't make the toast more native feeling but just adds another 
option.